Refactor loop indices and type handling for improved safety and consistency
- Replaced `int` with `size_t` in various loops across `CShareData.cpp`, `CDlgFuncList.cpp`, and `CPropCommon.cpp`, enhancing compatibility with larger datasets and ensuring safer memory management.
- Continued to document type usage distinctions in AGENTS.md, providing clear guidelines for new code and reinforcing overall type consistency across the codebase.

+++ sakura_core/outline/CDlgFuncList.cpp
@@ -1528,7 +1528,7 @@ void CDlgFuncList::SetTreeFile()
 	int nFuncInfo = 0;
 	std::vector<HTREEITEM> hParentTree;
 	hParentTree.push_back(TVI_ROOT);
-	for( int i = 0; i < (int)m_fileTreeSetting.m_aItems.size(); i++ ){
+	for( size_t i = 0; i < m_fileTreeSetting.m_aItems.size(); i++ ){
 		WCHAR szPath[_MAX_PATH];
 		WCHAR szPath2[_MAX_PATH];
 		const SFileTreeItem& item = m_fileTreeSetting.m_aItems[i];
